#7 split into domain folders
