+++ src/Auth.Wiedersehen/Configuration/ConfigurationKey.cs
@@ -6,4 +6,11 @@ internal struct ConfigurationKey
     {
         public const string MinLength = "Password:MinLength";
     }
+    
+    public struct ConnectionString
+    {
+        public const string ApplicationDb = "ApplicationDB";
+        public const string ConfigurationDb = "ConfigurationDB";
+        public const string PersistentGrandDb = "PersistentGrandDB";
+    }
 }
