+++ src/Auth.Wiedersehen/Exceptions/ErrorDetails.cs
@@ -1,3 +1,3 @@
-namespace Auth.Wiedersehen.Controllers.Models;
+namespace Auth.Wiedersehen.Exceptions;
 
 public record ErrorDetails(int StatusCode, IEnumerable<KeyValuePair<string, string>>? Errors);
