+++ src/Auth.Wiedersehen/Exceptions/HttpResponseExceptionFilter.cs
@@ -1,4 +1,3 @@
-using Auth.Wiedersehen.Controllers.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
